fixes and added additional blocks for pushdowns for monetdb

+++ plugins/monetdb-adapter/src/main/java/org/polypheny/db/adapter/monetdb/stores/MonetdbStore.java
@@ -306,7 +306,7 @@ public class MonetdbStore extends AbstractJdbcStore {
         return switch ( type ) {
             case BOOLEAN -> "BOOLEAN";
             case VARBINARY -> "VARCHAR";//throw new GenericRuntimeException( "Unsupported datatype: " + type.name() );
-            case TINYINT -> "TINYINT";
+            case TINYINT -> "SMALLINT"; // there seems to be an issue with tinyints and the jdbc driver
             case SMALLINT -> "SMALLINT";
             case INTEGER -> "INT";
             case BIGINT -> "BIGINT";
